requirements.txt
langgraph
langgraph-prebuilt
langgraph-sdk
langgraph-checkpoint-sqlite
langsmith
langchain-community
langchain-core
langchain-huggingface
langchain-ollama
langchain-text-splitters
langchain-chromadb
streamlit
langsmith
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
python-dotenv   
uuid_utils
pydantic
pydantic_core
langchain-tavily
PyMuPDF
pymupdf-layout
pymupdf4llm
langgraph-cli[inmem]
